Rename longrides test classes
