+++ long-ride-alerts/src/test/scala/org/apache/flink/training/exercises/longrides/scala/LongRidesIntegrationTest.scala
@@ -18,22 +18,24 @@
 
 package org.apache.flink.training.exercises.longrides.scala
 
+import org.apache.flink.streaming.api.functions.source.SourceFunction
 import org.apache.flink.training.exercises.common.datatypes.TaxiRide
 import org.apache.flink.training.exercises.longrides
-import org.apache.flink.training.exercises.testing.ComposedKeyedProcessFunction
-import org.apache.flink.training.solutions.longrides.scala.LongRidesSolution
+import org.apache.flink.training.exercises.testing.{ComposedPipeline, ExecutablePipeline, TestSink}
 
 /**
-  * The Scala tests extend the Java tests by overriding the composedAlertFunction() method
+  * The Scala tests extend the Java tests by overriding the longRidesPipeline() method
   * to use the Scala implementations of the exercise and solution.
   */
-class LongRidesHarnessTest extends longrides.LongRidesHarnessTest {
+class LongRidesIntegrationTest extends longrides.LongRidesIntegrationTest {
+  private val EXERCISE: ExecutablePipeline[TaxiRide, Long] =
+    (source: SourceFunction[TaxiRide], sink: TestSink[Long]) =>
+      (new LongRidesExercise.LongRidesJob(source, sink)).execute()
 
-  private val scalaExercise = new LongRidesExercise.AlertFunction
+  private val SOLUTION: ExecutablePipeline[TaxiRide, Long] =
+    (source: SourceFunction[TaxiRide], sink: TestSink[Long]) =>
+      (new LongRidesSolution.LongRidesJob(source, sink)).execute()
 
-  private val scalaSolution = new LongRidesSolution.AlertFunction
-
-  override def composedAlertFunction: ComposedKeyedProcessFunction[Long, TaxiRide, Long] =
-    new ComposedKeyedProcessFunction[Long, TaxiRide, Long](scalaExercise, scalaSolution)
-
+  override def longRidesPipeline: ComposedPipeline[TaxiRide, Long] =
+    new ComposedPipeline[TaxiRide, Long](EXERCISE, SOLUTION)
 }
